fix: move common asset types from queries/assets.rs to types.rs
