+++ modules/rest_blockfrost/src/types.rs
@@ -2,12 +2,9 @@ use crate::cost_models::{PLUTUS_V1, PLUTUS_V2, PLUTUS_V3};
 use acropolis_common::{
     messages::EpochActivityMessage,
     protocol_params::{Nonce, NonceVariant, ProtocolParams},
-    queries::{
-        assets::{AssetMetadataStandard, AssetMintRecord, PolicyAsset},
-        governance::DRepActionUpdate,
-    },
+    queries::governance::DRepActionUpdate,
     rest_helper::ToCheckedF64,
-    PoolEpochState, Relay, Vote,
+    AssetMetadataStandard, AssetMintRecord, PolicyAsset, PoolEpochState, Relay, Vote,
 };
 use num_traits::ToPrimitive;
 use rust_decimal::Decimal;
